app/requirements.txt
streamlit
langchain-text-splitters==0.3.4
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1
sentence-transformers==3.3.1